xLARFP reflects a vector to a non-negative value.
The code is made stable by algebraically expanding the result and
cancelling equal terms in the BETA.GE.ZERO branch.  Complex needs
a bit of care; an extra imaginary term appears from the conjugate
multiplication.  The algebra and choices will be provided in a
separate LAWN-ish document, but this code should be as safe for
over- and underflow as xLARFG.

Signed-off-by: Jason Riedy <ejr@cs.berkeley.edu>