This is the codes for the new DQDS of Osni Marques:    DLASQ1, DLASQ2, DLASQ3, DLASQ4, DLASQ5, DLASQ6,    SLASQ1, SLASQ2, SLASQ3, SLASQ4, SLASQ5, SLASQ6.
There are interface changes for:
  DLASQ3, DLASQ4, SLASQ3, SLASQ4.
since we decided *not* to add another set of routines and simply
change the interfaces.

I have removed Sven's code 
  DLAZQ3, DLAZQ4, SLAZQ3, SLAZQ4.
They were introduced in 3.1 for thread safety but are no longer needed.

This code (3.2) passes Cleve Moler's matrix while 3.0 and 3.1 do not pass this
matrix.  The algorithm in 3.0 and 3.1 fails to converge in the maximum number
of iterations allowed. Cleve Moler's matrix is at:
http://www-math.cudenver.edu/~langou/lapack-3.2/set_Moler_200.c

The single precision version of the code (3.2) has a problem in the testing 
with one matrix of type 16. (3.0 and 3.1 were fine). I have extracted the
matrix, you can get it:
http://www-math.cudenver.edu/~langou/lapack-3.2/set_PB1_30.c
To pass this matrix, we needed to set the IEEE flag in SLASQ2 to .FALSE..
This is a temporary fix, we wish to fix this better in the future.

A last note, this is not exactly Osni's code, I have:
* change the header
* remove some blank space at the end of the lines
* add Jason's patch: replace the 
     DMIN.NE.DMIN 
  (use to detect NaNs) by
      DISNAN( DMIN ) in DLASQ3
  and SISNAN( DMIN ) in SLASQ3.