./lapack/bugs/bug112
Adding the program for bug112. 

The program was created by Daniel Strobusch.

This is a fairly simple program but we could reuse it to test some potential
thread safe issues with LAPACK. (So I put it here for further references.) In
this case, the program creates M symmetric N-by-N matrices to be diagonalized
in parallel using MRRR. The program uses OpenMP parallel for. Nothing fancy but
good to have under the hand.