correcting bug 128 See: http://icl.cs.utk.edu/lapack-forum/viewtopic.php?t=4746
When LAPACK_NAN_CHECK is enabled, in input of xLASET, we now allow if the matrix A contains NaN. 
 
Since the goal of xLASET is to set values in A, we consider it OK if A contains NaN in input.

In other words, the rationale is that we do not consider A has an input/output. We consider A as an output only.

Note: We allow neither alpha nor beta to be a NaN.

(So we do check for NaN in alpha and in beta and return error if NaN.)

So this commit
(1) removes the NaN check on matrix A in LAPACKE wrappers: lapacke_?laset.c
(2) declares that A is OUTPUT in LAPACK subroutines: ?laset.f
(3) adds a thanks to Victor Kostin from Intel