Patch from Andreas Noack (MIT).
The code for DGEEV was checking the INFO parameter returned by DHSEQR as
    IF( INFO.GT.0 )THEN "QUIT"

Andreas proposed to check
    IF( INFO.NE.0 )THEN "QUIT"

I am fine with the modification. Not terribly excited but fine.

There is really no reason for the case ( INFO.LT.0 ) to happen, so we should
not have to deal with this case.  If this happen, it is really a bug and we
should be notified.  But, at this point, why not.

I applied the proposed patch to xGEEV and xGEEVX.

Julien.